agents-workshop/01_create_tools/requirements.txt
pydantic
mlflow
unitycatalog-langchain[databricks]
unitycatalog-ai[databricks]
pyyaml
tabulate
pandas